AO/InitialExperiments/requirements.txt
autograd==1.5
cycler==0.11.0
decorator==4.4.2
future==0.18.3
importlib-resources==5.4.0
kiwisolver==1.3.1
matplotlib==3.3.4
networkx==2.5.1
numpy==1.19.5
pandas==1.1.5
Pillow==8.4.0
pymanopt==2.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
scipy==1.5.4
six==1.16.0
tqdm==4.64.1
zipp==3.6.0
